% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postalcodes.R
\name{cleanPostalCodes}
\alias{cleanPostalCodes}
\title{Clean postal codes}
\usage{
cleanPostalCodes(codes, blank = TRUE)
}
\arguments{
\item{codes}{vector that is coercible to a character vector of
postal codes}

\item{blank}{should there be a blank space between the two parts
of the code?}
}
\description{
Clean postal codes
}
\examples{
cds <- c("A1B 2C3", "missing",
         "---  A    5G*****HHH5J4",
         "E9J6H2nnnn", "e6n8l0")
cleanPostalCodes(cds, TRUE)
cleanPostalCodes(cds, FALSE)
cleanPostalCodes(cds, TRUE, "missing")
cleanPostalCodes(cds, FALSE, "missing")
}

